+++ epcis_event_hash_generator/__main__.py
@@ -33,7 +33,7 @@ import sys
 from epcis_event_hash_generator import hash_generator, events_from_file_reader
 
 
-def epcis_hash_from_file(path, hashalg="sha256", enforce="", join_by=""):
+def epcis_hash_from_file(path, hashalg="sha256", enforce="", join_by="", cbv_version="CBV2.0"):
     """
     This method exemplifies how to read all EPCIS Events from the EPCIS document in the file at path.
     The file is parsed extracting the events data. The pre hash string is computed for each event.
@@ -43,7 +43,7 @@ def epcis_hash_from_file(path, hashalg="sha256", enforce="", join_by=""):
     events = events_from_file_reader.event_list_from_file(path, enforce)
 
     prehashes = hash_generator.derive_prehashes_from_events(events, join_by)
-    hashes = hash_generator.calculate_hashes_from_pre_hashes(prehashes, hashalg)
+    hashes = hash_generator.calculate_hashes_from_pre_hashes(prehashes, hashalg, cbv_version)
 
     return hashes, prehashes
 
@@ -93,6 +93,12 @@ def command_line_parsing():
         + " Defaults to guessing the format from the file ending.",
         choices=["XML", "JSON", ""],
         default="")
+    parser.add_argument(
+        "-v",
+        "--version",
+        help="CBV version to use in hash generation. Defaults to CBV2.0.",
+        choices=["CBV2.0", "CBV2.1"],
+        default="CBV2.0")
 
     args = parser.parse_args()
 
@@ -124,7 +130,7 @@ def main():
     for filename in args.file:
         # ACTUAL ALGORITHM CALL:
         (hashes, prehashes) = epcis_hash_from_file(
-            path=filename, hashalg=args.algorithm, join_by=args.join, enforce=args.enforce_format)
+            path=filename, hashalg=args.algorithm, join_by=args.join, enforce=args.enforce_format, cbv_version=args.version)
 
         # Output:
         if args.batch: